requirements.txt
django==2.2.7
django-bootstrap4==1.0.1
django-tables2==2.1.1
djangorestframework==3.10.3
markdown==2.6.11
django-filter==2.2.0
coreapi==2.3.3
